Wrapping up ViewDecorator, and to clear up some confusion
Three new signals were added:

- "detached" -> DetachedSignalData
- "swf-deco-fini" -> DecoratorFinishSinal
- "swf-deco-config-changed" -> ConfigChangedSignal

However, it should be noted that ViewDecorator has a brand new
signal. This is because before, we used "signal strings" to designate
signals, and multiple "signal strings" could call the same
handle.

Before, both the "swf-deco-fini" and the "detached" signal would
trigger the on_detached callback.

Now, we don't have "signal strings", the dispatch appears to be some
template black magic to dispatch on the type alone. So, I'm adding an
extra callback that does the same thing